Move p2p to net, update implementation
Refactor the implementation to implement the TransportIdentifier
interface as well as modifying methods to use said interface (as well as
unexporting methods). Remove the need for net.TransportIdentifier.ID().
