+++ pkg/net/libp2p.go
@@ -1,4 +1,4 @@
-package p2p
+package net
 
 import (
 	"fmt"
@@ -10,46 +10,47 @@ import (
 	ci "github.com/libp2p/go-libp2p-crypto"
 	peer "github.com/libp2p/go-libp2p-peer"
 	pstore "github.com/libp2p/go-libp2p-peerstore"
-
-	"github.com/keep-network/keep-core/pkg/net"
 )
 
+// Implementation of the TransportIdentifier interface
+type networkID peer.ID
+
+func (ni networkID) ProviderName() string {
+	return "libp2p"
+}
+
+// peerIdentity represents a group member's network level identity. A valid group
+// member will generate or provide a keypair, which will correspond to a network
+// ID. Consumers of the net package require an ID to register with protocol level
+// ID's, as well as a public key for authentication.
 type peerIdentity struct {
 	privKey ci.PrivKey
 }
 
-func (pi *peerIdentity) ProviderName() string {
-	return "libp2p"
+func (pi *peerIdentity) ID() TransportIdentifier {
+	return networkID(pubKeyToID(pi.privKey.GetPublic()))
 }
 
-func (pi *peerIdentity) ID() string {
-	return pubKeyToID(pi.privKey.GetPublic())
-}
-
-func pubKeyToID(pub ci.PubKey) string {
+func pubKeyToID(pub ci.PubKey) peer.ID {
 	pid, err := peer.IDFromPublicKey(pub)
 	if err != nil {
 		panic(fmt.Sprintf("Failed to generate valid libp2p identity with err: %v", err))
 	}
-	return string(pid)
+	return pid
 }
 
-func (pi *peerIdentity) PubKeyFromID(id string) ([]byte, error) {
-	key, err := peer.ID(id).ExtractPublicKey()
-	if err != nil {
-		return nil, err
-	}
-	return key.Bytes()
+func (pi *peerIdentity) PubKeyFromID(id TransportIdentifier) (ci.PubKey, error) {
+	pid := peer.ID(id.(networkID))
+	return pid.ExtractPublicKey()
 }
 
 // AddIdentityToStore takes a peerIdentity and notifies the addressbook of the
 // existance of a new client joining the network.
-func AddIdentityToStore(pi *peerIdentity) (pstore.Peerstore, error) {
+func addIdentityToStore(pi *peerIdentity) (pstore.Peerstore, error) {
 	// TODO: investigate a generic store interface that gives us a unified interface
 	// to our address book (peerstore in libp2p) from secure storage (dht)
 	ps := pstore.NewPeerstore()
-
-	id := peer.ID(pi.ID())
+	id := peer.ID(pi.ID().(networkID))
 
 	if err := ps.AddPrivKey(id, pi.privKey); err != nil {
 		return nil, fmt.Errorf("failed to add PrivateKey with error %s", err)
@@ -60,9 +61,9 @@ func AddIdentityToStore(pi *peerIdentity) (pstore.Peerstore, error) {
 	return ps, nil
 }
 
-// LoadOrGenerateIdentity allows a client to provide or generate an Identity that
+// loadOrGenerateIdentity allows a client to provide or generate an Identity that
 // will be used to reference the client in the peer-to-peer network.
-func LoadOrGenerateIdentity(randseed int64, filePath string) (net.Identity, error) {
+func loadOrGenerateIdentity(randseed int64, filePath string) (*peerIdentity, error) {
 	if filePath != "" {
 		// TODO: unmarshal and build out PKI
 		// TODO: ensure this is associated with some staking address
@@ -76,7 +77,7 @@ func LoadOrGenerateIdentity(randseed int64, filePath string) (net.Identity, erro
 // generateIdentity generates a public/private-key pair
 // (using the libp2p/crypto wrapper for golang/crypto) provided a reader.
 // Use randseed for deterministic IDs, otherwise we'll use cryptographically secure psuedorandomness.
-func generateIdentity() (net.Identity, error) {
+func generateIdentity() (*peerIdentity, error) {
 	var r io.Reader
 	r = crand.Reader
 
@@ -88,7 +89,7 @@ func generateIdentity() (net.Identity, error) {
 	return &peerIdentity{privKey: priv}, nil
 }
 
-func generateDeterministicIdentity(randseed int64) (net.Identity, error) {
+func generateDeterministicIdentity(randseed int64) (*peerIdentity, error) {
 	var r io.Reader
 	r = mrand.New(mrand.NewSource(randseed))
